Complete the basic project modules.
Completed the main project modules including database,services
 and rest controllers. I added an exception helper class to handle
the logical exceptions in the controller layer at one place. I also added
internationlization for the exception handlers. I integrated my api with swagger
and documented both user controller and file controller.

The next step is to complete the swagger documentation for
the remaining controllers and to add junit testing to test
 the basic functionality of my api. Also I will use
 jmeter to make some test plans.